+++ source/isaaclab_tasks/isaaclab_tasks/direct/constrained_albc/mdp/__init__.py
@@ -46,6 +46,7 @@ from .rewards import (
     RewardTermCfg,
     action_smoothness_penalty,
     command_reward,
+    joint_torque_penalty,
 )
 
 __all__ = [
@@ -83,4 +84,5 @@ __all__ = [
     # Rewards (functions)
     "action_smoothness_penalty",
     "command_reward",
+    "joint_torque_penalty",
 ]